+++ R/play.R
@@ -4,9 +4,6 @@
 #'
 #' @param x Object to be played.
 #' @inheritParams pt2_render
-#' @param note Note to be played when `x` is a `pt2samp` class object. Defaults
-#' to `"C-2"`.
-#' @samples TODO
 #' @param ... Arguments passed to [`pt2_render()`].
 #' @returns Returns an `[audio::`$.audioInstance`]` object which
 #' allows you to control the playback (pause, resume, rewind).
@@ -43,27 +40,30 @@ play.pt2mod <- function(x, duration = NA, options = pt2_render_options(), positi
 #' @name play
 #' @rdname play
 #' @method play pt2samp
-play.pt2samp <- function(x, duration = 5, options = pt2_render_options(), note = "C-2", ...) {
-  mod <- pt2_new_mod("play.samp")
-  mod$samples[[1]] <- x
-  mod$patterns[[1]][1,c(1,2)][[1]] <- paste0(note, " 01 F00")
-  play(x = mod, options = options, duration = duration, ...)
+play.pt2samp <- function(x, duration = 5, options = pt2_render_options(), note = "C-3", ...) {
+  x <- pt2_render(x, duration = duration, options = options, note = note, ...)
+  rate <- attributes(x)$rate
+  audio::play.audioSample(x, rate = rate)
 }
 
 #' @export
 #' @name play
 #' @rdname play
 #' @method play pt2patlist
-play.pt2patlist <- function(x, duration = 5, options = pt2_render_options(), samples, ...) {
-  #TODO
+play.pt2patlist <- function(x, duration = NA, options = pt2_render_options(), samples, ...) {
+  x <- pt2_render(x, duration = duration, options = options, samples = samples, ...)
+  rate <- attributes(x)$rate
+  audio::play.audioSample(x, rate = rate)
 }
 
 #' @export
 #' @name play
 #' @rdname play
 #' @method play pt2pat
-play.pt2pat <- function(x, duration = 5, options = pt2_render_options(), samples, ...) {
-  #TODO
+play.pt2pat <- function(x, duration = NA, options = pt2_render_options(), samples, ...) {
+  x <- pt2_render(x, duration = duration, options = options, samples = samples, ...)
+  rate <- attributes(x)$rate
+  audio::play.audioSample(x, rate = rate)
 }
 
 #' @export
@@ -71,7 +71,9 @@ play.pt2pat <- function(x, duration = 5, options = pt2_render_options(), samples
 #' @rdname play
 #' @method play pt2celllist
 play.pt2celllist <- function(x, duration = 5, options = pt2_render_options(), samples, ...) {
-  #TODO
+  x <- pt2_render(x, duration = duration, options = options, samples = samples, ...)
+  rate <- attributes(x)$rate
+  audio::play.audioSample(x, rate = rate)
 }
 
 #' @export
@@ -79,5 +81,7 @@ play.pt2celllist <- function(x, duration = 5, options = pt2_render_options(), sa
 #' @rdname play
 #' @method play pt2cell
 play.pt2cell <- function(x, duration = 5, options = pt2_render_options(), samples, ...) {
-  #TODO
+  x <- pt2_render(x, duration = duration, options = options, samples = samples, ...)
+  rate <- attributes(x)$rate
+  audio::play.audioSample(x, rate = rate)
 }