+++ src/mod_header.h
@@ -5,4 +5,4 @@
 using namespace cpp11;
 
 int n_patterns_internal(SEXP mod);
-void set_mod_name_internal(module_t * my_song, strings name);
+void set_mod_name_internal(module_t * my_song, std::string name);